int -> size_t

+++ inc/c_array.h
@@ -26,24 +26,24 @@
 #define _C_ARRAY_H_
 
 struct cstl_array {
-    int capacity; /* Number of maximum elements array can hold without reallocation */
-    int count;  /* Number of current elements in the array */
+    size_t capacity; /* Number of maximum elements array can hold without reallocation */
+    size_t count;  /* Number of current elements in the array */
     struct cstl_object** pElements; /* actual storage area */
     cstl_compare compare_fn; /* Compare function pointer*/
     cstl_destroy destruct_fn; /* Destructor function pointer*/
 };
 
-extern struct cstl_array* cstl_array_new ( int init_size, cstl_compare fn_c, cstl_destroy fn_d);
+extern struct cstl_array* cstl_array_new ( size_t init_size, cstl_compare fn_c, cstl_destroy fn_d);
 extern cstl_error cstl_array_push_back ( struct cstl_array* pArray, void* elem, size_t elem_size);
-extern const void * cstl_array_element_at(struct cstl_array* pArray, int index);
-extern cstl_error cstl_array_insert_at ( struct cstl_array* pArray, int index, void* elem, size_t elem_size);
-extern int cstl_array_size( struct cstl_array* pArray);
-extern int cstl_array_capacity( struct cstl_array* pArray );
+extern const void * cstl_array_element_at(struct cstl_array* pArray, size_t index);
+extern cstl_error cstl_array_insert_at ( struct cstl_array* pArray, size_t index, void* elem, size_t elem_size);
+extern size_t cstl_array_size( struct cstl_array* pArray);
+extern size_t cstl_array_capacity( struct cstl_array* pArray );
 extern cstl_bool  cstl_array_empty( struct cstl_array* pArray);
-extern cstl_error cstl_array_reserve( struct cstl_array* pArray, int pos);
+extern cstl_error cstl_array_reserve( struct cstl_array* pArray, size_t pos);
 extern const void * cstl_array_front(struct cstl_array* pArray);
 extern const void * cstl_array_back(struct cstl_array* pArray);
-extern cstl_error cstl_array_remove_from ( struct cstl_array*, int pos);
+extern cstl_error cstl_array_remove_from ( struct cstl_array*, size_t pos);
 extern cstl_error cstl_array_delete( struct cstl_array* pArray);
 
 extern struct cstl_iterator* cstl_array_new_iterator(struct cstl_array* pArray);
